requirements.txt
beautifulsoup4
colorama
curl-cffi
fake-useragent
names
requests
2captcha-python
anticaptchaofficial
python-dotenv
